refactor(annotation-server): organize interfaces
