+++ annotation-server/src/medications/medication.entity.ts
@@ -1,23 +1,6 @@
 import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
 
-interface ExternalIdentifier {
-    resource: 'PharmGKB' | string;
-    identifier: string;
-}
-interface InternationalBrand {
-    name: string;
-}
-
-export interface Drug {
-    name: string;
-    description?: string;
-    'external-identifiers'?: {
-        'external-identifier': Array<ExternalIdentifier> | ExternalIdentifier;
-    };
-    'international-brands'?: {
-        'international-brand': Array<InternationalBrand> | InternationalBrand;
-    };
-}
+import { Drug } from './interfaces/drugbank.interface';
 
 @Entity()
 export class Medication {
